move exports around.
